feat: add PersonSidebar component and update related imports; refactor component usage in various routes
