+++ apps/next-web/src/routes/fields/+page.svelte
@@ -2,7 +2,7 @@
   import cytoscape, { type NodeSingular } from "cytoscape";
   import { onMount } from "svelte";
   import type { PageProps } from "./$types";
-  import PersonSidebar from "./PersonSidebar.svelte";
+  import PersonSidebar from "../../lib/components/PersonSidebar.svelte";
   import { style } from "./cytoscape";
   import ToggleButton from "$lib/components/ToggleButton.svelte";
