+++ apps/next-web/src/routes/human/LifeSpan.svelte
@@ -39,42 +39,6 @@
 <article>
   {lifespan_properties.size}
 </article>
-<div class="llm">
-  <blockquote class="twitter-tweet" data-dnt="true">
-    <p lang="en" dir="ltr">
-      Instead of watching an hour of Netflix, watch this 2 hour hour Stanford
-      lecture will teach you more about how LLMs like ChatGPT and Claude are
-      built than most people working at top AI companies learn in their entire
-      careers. <a href="https://t.co/Htfn54LtDG">pic.twitter.com/Htfn54LtDG</a>
-    </p>
-    &mdash; AilaunchX (@Ai_Tech_tool)<a
-      href="https://twitter.com/Ai_Tech_tool/status/2054595670854770933?ref_src=twsrc%5Etfw"
-      >May 13, 2026</a
-    >
-  </blockquote>
-  <script
-    async
-    src="https://platform.twitter.com/widgets.js"
-    charset="utf-8"
-  ></script>
-</div>
-<div class="supps">
-  <blockquote class="twitter-tweet">
-    <p lang="en" dir="ltr">
-      The only 5 supplements worth your money (and why each one matters):<br
-      /><br />1. Vitamin D3 + K2.
-    </p>
-    &mdash; Leddy (@LeddyLLC)<a
-      href="https://twitter.com/LeddyLLC/status/2054921528790426090?ref_src=twsrc%5Etfw"
-      >May 14, 2026</a
-    >
-  </blockquote>
-  <script
-    async
-    src="https://platform.twitter.com/widgets.js"
-    charset="utf-8"
-  ></script>
-</div>
 
 <style>
   article {
